sleep-disorder/requirements.txt
confluent-kafka==2.6.1
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.1
kafka-python==2.0.2
kiwisolver==1.4.7
matplotlib==3.9.3
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
py4j==0.10.9.7
pyparsing==3.2.0
pyspark==3.5.3
python-dateutil==2.9.0.post0
pytz==2024.2
setuptools==75.6.0
six==1.17.0
tabulate==0.9.0
tzdata==2024.2
